pyproject.toml
[project]
name            = "genie"
version         = "0.1.0"
authors = [
  {name = "Carlos Fernandez Musoles", email = "carlos@kalavai.net"}
]
maintainers = [
  {name = "Carlos Fernandez Musoles", email = "carlos@kalavai.net"}
]
description     = "Hackathon experiment for agentic GenAI Ops"
license         = "Apache-2.0"
license-files   = ["LICENSE"]
keywords        = ["LLM", "platform"]
requires-python = ">=3.4"
dependencies = [
    "requests>= 2.25",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "streamlit>=1.32.0",
    "reflex>=0.7.0",
    "llama-index>=0.9.0",
    "llama-index-llms-openai-like",
    "streamlit-audiorecorder",
    "httpx"
]


[project.urls]
Homepage        = "https://platform.kalavai.net"
Website         = "https://kalavai.net"


[build-system]
requires        = [
    "poetry-core>=1.0.0",
    "setuptools"  # for pip install -e
]
build-backend   = "poetry.core.masonry.api"



[tool.setuptools.packages.find]
where = ["tools"]

[tool.setuptools.package-data]
assets = ["*.yaml", "*.conf"]